--- src/main/java/fr/abes/thesesapirecherche/dto/OrganismeResponseDto.java
@@ -1,13 +1,0 @@
-package fr.abes.thesesapirecherche.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Builder
-@Getter
-@Setter
-public class OrganismeResponseDto {
-    private String ppn;
-    private String nom;
-}
