--- src/main/java/fr/abes/thesesapirecherche/dto/ThesePersoneResponseDto.java
@@ -1,14 +1,0 @@
-package fr.abes.thesesapirecherche.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Builder
-@Getter
-@Setter
-public class ThesePersoneResponseDto {
-    private String ppn;
-    private String nom;
-    private String prenom;
-}
