--- src/main/java/fr/abes/thesesapirecherche/dto/TheseResponseDto.java
@@ -1,30 +1,0 @@
-package fr.abes.thesesapirecherche.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-import java.util.Map;
-
-@Builder
-@Getter
-@Setter
-public class TheseResponseDto {
-
-    String titrePrincipal;
-    String nnt;
-    String dateSoutenance;
-    String discipline;
-    Map<String, String> titres;
-    Map<String, String> resumes;
-    OrganismeResponseDto etabSoutenance;
-    List<OrganismeResponseDto> etabCotutelle;
-    List<OrganismeResponseDto> partenairesRecherche;
-    List<String> sujetsRameau;
-    List<ThesePersoneResponseDto> membresJury;
-    List<ThesePersoneResponseDto> rapporteurs;
-    List<ThesePersoneResponseDto> auteurs;
-    List<ThesePersoneResponseDto> directeurs;
-
-}
