--- src/main/java/fr/abes/thesesapirecherche/model/Organisme.java
@@ -1,11 +1,0 @@
-package fr.abes.thesesapirecherche.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class Organisme {
-    String ppn;
-    String nom;
-}
