--- src/main/java/fr/abes/thesesapirecherche/model/PersonneThese.java
@@ -1,16 +1,0 @@
-package fr.abes.thesesapirecherche.model;
-
-
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * Représente une personne dans une thèse
- */
-@Getter
-@Setter
-public class PersonneThese {
-    private String ppn;
-    private String nom;
-    private String prenom;
-}
