--- src/main/java/fr/abes/thesesapirecherche/model/ThesePersonne.java
@@ -1,19 +1,0 @@
-package fr.abes.thesesapirecherche.model;
-
-import lombok.*;
-
-/**
- * Représente une thèse pour une personne
- */
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ThesePersonne {
-
-    @Getter @Setter
-    String nnt;
-
-    @Getter @Setter
-    String role;
-
-}
