+++ src/main/java/fr/abes/thesesapirecherche/theses/dto/OrganismeResponseDto.java
@@ -1,0 +1,13 @@
+package fr.abes.thesesapirecherche.theses.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Builder
+@Getter
+@Setter
+public class OrganismeResponseDto {
+    private String ppn;
+    private String nom;
+}
