+++ src/main/java/fr/abes/thesesapirecherche/theses/model/Organisme.java
@@ -1,0 +1,11 @@
+package fr.abes.thesesapirecherche.theses.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Organisme {
+    String ppn;
+    String nom;
+}
